Moved all remaining tests in test_common to dedicated files
