+++ tests/expr_and_series/binary_test.py
@@ -1,0 +1,48 @@
+from typing import Any
+import pytest
+
+import narwhals.stable.v1 as nw
+from tests.utils import compare_dicts
+
+
+def test_expr_binary(constructor: Any, request: Any) -> None:
+    data = {"a": [1, 3, 2], "b": [4, 4, 6], "z": [7.0, 8, 9]}
+    if "dask" in str(constructor):
+        request.applymarker(pytest.mark.xfail)
+    df_raw = constructor(data)
+    result = nw.from_native(df_raw).with_columns(
+        a=(1 + 3 * nw.col("a")) * (1 / nw.col("a")),
+        b=nw.col("z") / (2 - nw.col("b")),
+        c=nw.col("a") + nw.col("b") / 2,
+        d=nw.col("a") - nw.col("b"),
+        e=((nw.col("a") > nw.col("b")) & (nw.col("a") >= nw.col("z"))).cast(nw.Int64),
+        f=(
+            (nw.col("a") < nw.col("b"))
+            | (nw.col("a") <= nw.col("z"))
+            | (nw.col("a") == 1)
+        ).cast(nw.Int64),
+        g=nw.col("a") != 1,
+        h=(False & (nw.col("a") != 1)),
+        i=(False | (nw.col("a") != 1)),
+        j=2 ** nw.col("a"),
+        k=2 // nw.col("a"),
+        l=nw.col("a") // 2,
+        m=nw.col("a") ** 2,
+    )
+    expected = {
+        "a": [4, 3.333333, 3.5],
+        "b": [-3.5, -4.0, -2.25],
+        "z": [7.0, 8.0, 9.0],
+        "c": [3, 5, 5],
+        "d": [-3, -1, -4],
+        "e": [0, 0, 0],
+        "f": [1, 1, 1],
+        "g": [False, True, True],
+        "h": [False, False, False],
+        "i": [False, True, True],
+        "j": [2, 8, 4],
+        "k": [2, 0, 1],
+        "l": [0, 1, 1],
+        "m": [1, 9, 4],
+    }
+    compare_dicts(result, expected)
