+++ tests/frame/invalid_test.py
@@ -1,0 +1,20 @@
+import pytest
+import polars as pl
+import pandas as pd
+import pyarrow as pa
+
+import narwhals.stable.v1 as nw
+
+
+def test_invalid() -> None:
+    data = {"a": [1, 3, 2], "b": [4, 4, 6], "z": [7.0, 8, 9]}
+    df = nw.from_native(pa.table({"a": [1, 2], "b": [3, 4]}))
+    with pytest.raises(ValueError, match="Multi-output"):
+        df.select(nw.all() + nw.all())
+    df = nw.from_native(pd.DataFrame(data))
+    with pytest.raises(ValueError, match="Multi-output"):
+        df.select(nw.all() + nw.all())
+    with pytest.raises(TypeError, match="Perhaps you:"):
+        df.select([pl.col("a")])  # type: ignore[list-item]
+    with pytest.raises(TypeError, match="Perhaps you:"):
+        df.select([nw.col("a").cast(pl.Int64)])
