--- tests/frame/test_common.py
@@ -1,140 +1,0 @@
-from __future__ import annotations
-
-from typing import Any
-
-import pandas as pd
-import polars as pl
-import pyarrow as pa
-import pytest
-
-import narwhals.stable.v1 as nw
-from tests.utils import compare_dicts
-
-data = {"a": [1, 3, 2], "b": [4, 4, 6], "z": [7.0, 8, 9]}
-data_na = {"a": [None, 3, 2], "b": [4, 4, 6], "z": [7.0, None, 9]}
-data_right = {"c": [6, 12, -1], "d": [0, -4, 2]}
-
-
-@pytest.mark.filterwarnings("ignore:Determining|Resolving.*")
-def test_columns(constructor: Any) -> None:
-    df = nw.from_native(constructor(data))
-    result = df.columns
-    expected = ["a", "b", "z"]
-    assert result == expected
-
-
-def test_expr_binary(constructor: Any, request: Any) -> None:
-    if "dask" in str(constructor):
-        request.applymarker(pytest.mark.xfail)
-    df_raw = constructor(data)
-    result = nw.from_native(df_raw).with_columns(
-        a=(1 + 3 * nw.col("a")) * (1 / nw.col("a")),
-        b=nw.col("z") / (2 - nw.col("b")),
-        c=nw.col("a") + nw.col("b") / 2,
-        d=nw.col("a") - nw.col("b"),
-        e=((nw.col("a") > nw.col("b")) & (nw.col("a") >= nw.col("z"))).cast(nw.Int64),
-        f=(
-            (nw.col("a") < nw.col("b"))
-            | (nw.col("a") <= nw.col("z"))
-            | (nw.col("a") == 1)
-        ).cast(nw.Int64),
-        g=nw.col("a") != 1,
-        h=(False & (nw.col("a") != 1)),
-        i=(False | (nw.col("a") != 1)),
-        j=2 ** nw.col("a"),
-        k=2 // nw.col("a"),
-        l=nw.col("a") // 2,
-        m=nw.col("a") ** 2,
-    )
-    expected = {
-        "a": [4, 3.333333, 3.5],
-        "b": [-3.5, -4.0, -2.25],
-        "z": [7.0, 8.0, 9.0],
-        "c": [3, 5, 5],
-        "d": [-3, -1, -4],
-        "e": [0, 0, 0],
-        "f": [1, 1, 1],
-        "g": [False, True, True],
-        "h": [False, False, False],
-        "i": [False, True, True],
-        "j": [2, 8, 4],
-        "k": [2, 0, 1],
-        "l": [0, 1, 1],
-        "m": [1, 9, 4],
-    }
-    compare_dicts(result, expected)
-
-
-def test_expr_transform(constructor: Any, request: Any) -> None:
-    if "dask" in str(constructor):
-        request.applymarker(pytest.mark.xfail)
-    df = nw.from_native(constructor(data))
-    result = df.with_columns(a=nw.col("a").is_between(-1, 1), b=nw.col("b").is_in([4, 5]))
-    expected = {"a": [True, False, False], "b": [True, True, False], "z": [7, 8, 9]}
-    compare_dicts(result, expected)
-
-
-def test_expr_na(constructor: Any, request: Any) -> None:
-    if "dask" in str(constructor):
-        request.applymarker(pytest.mark.xfail)
-    df = nw.from_native(constructor(data_na)).lazy()
-    result_nna = df.filter((~nw.col("a").is_null()) & (~df.collect()["z"].is_null()))
-    expected = {"a": [2], "b": [6], "z": [9]}
-    compare_dicts(result_nna, expected)
-
-
-def test_lazy(constructor_eager: Any) -> None:
-    df = nw.from_native(constructor_eager({"a": [1, 2, 3]}), eager_only=True)
-    result = df.lazy()
-    assert isinstance(result, nw.LazyFrame)
-
-
-def test_invalid() -> None:
-    df = nw.from_native(pa.table({"a": [1, 2], "b": [3, 4]}))
-    with pytest.raises(ValueError, match="Multi-output"):
-        df.select(nw.all() + nw.all())
-    df = nw.from_native(pd.DataFrame(data))
-    with pytest.raises(ValueError, match="Multi-output"):
-        df.select(nw.all() + nw.all())
-    with pytest.raises(TypeError, match="Perhaps you:"):
-        df.select([pl.col("a")])  # type: ignore[list-item]
-    with pytest.raises(TypeError, match="Perhaps you:"):
-        df.select([nw.col("a").cast(pl.Int64)])
-
-
-@pytest.mark.parametrize("df_raw", [pd.DataFrame(data)])
-def test_reindex(df_raw: Any) -> None:
-    df = nw.from_native(df_raw, eager_only=True)
-    result = df.select("b", df["a"].sort(descending=True))
-    expected = {"b": [4, 4, 6], "a": [3, 2, 1]}
-    compare_dicts(result, expected)
-    result = df.select("b", nw.col("a").sort(descending=True))
-    compare_dicts(result, expected)
-
-    s = df["a"]
-    result_s = s > s.sort()
-    assert not result_s[0]
-    assert result_s[1]
-    assert not result_s[2]
-    result = df.with_columns(s.sort())
-    expected = {"a": [1, 2, 3], "b": [4, 4, 6], "z": [7.0, 8.0, 9.0]}  # type: ignore[list-item]
-    compare_dicts(result, expected)
-    with pytest.raises(ValueError, match="Multi-output expressions are not supported"):
-        nw.to_native(df.with_columns(nw.all() + nw.all()))
-
-
-def test_with_columns_order(constructor_eager: Any) -> None:
-    df = nw.from_native(constructor_eager(data))
-    result = df.with_columns(nw.col("a") + 1, d=nw.col("a") - 1)
-    assert result.columns == ["a", "b", "z", "d"]
-    expected = {"a": [2, 4, 3], "b": [4, 4, 6], "z": [7.0, 8, 9], "d": [0, 2, 1]}
-    compare_dicts(result, expected)
-
-
-def test_with_columns_order_single_row(constructor_eager: Any) -> None:
-    df = nw.from_native(constructor_eager(data)[:1])
-    assert len(df) == 1
-    result = df.with_columns(nw.col("a") + 1, d=nw.col("a") - 1)
-    assert result.columns == ["a", "b", "z", "d"]
-    expected = {"a": [2], "b": [4], "z": [7.0], "d": [0]}
-    compare_dicts(result, expected)
