+++ tests/frame/with_columns_test.py
@@ -1,7 +1,10 @@
+from typing import Any
+
 import numpy as np
 import pandas as pd
 
-import narwhals as nw
+import narwhals.stable.v1 as nw
+from tests.utils import compare_dicts
 
 
 def test_with_columns_int_col_name_pandas() -> None:
@@ -13,3 +16,22 @@ def test_with_columns_int_col_name_pandas() -> None:
         {0: [1, 4, 7], 1: [2, 5, 8], 2: [3, 6, 9], 4: [2, 5, 8]}, dtype="int64"
     )
     pd.testing.assert_frame_equal(result, expected)
+
+
+def test_with_columns_order(constructor_eager: Any) -> None:
+    data = {"a": [1, 3, 2], "b": [4, 4, 6], "z": [7.0, 8, 9]}
+    df = nw.from_native(constructor_eager(data))
+    result = df.with_columns(nw.col("a") + 1, d=nw.col("a") - 1)
+    assert result.columns == ["a", "b", "z", "d"]
+    expected = {"a": [2, 4, 3], "b": [4, 4, 6], "z": [7.0, 8, 9], "d": [0, 2, 1]}
+    compare_dicts(result, expected)
+
+
+def test_with_columns_order_single_row(constructor_eager: Any) -> None:
+    data = {"a": [1, 3, 2], "b": [4, 4, 6], "z": [7.0, 8, 9]}
+    df = nw.from_native(constructor_eager(data)[:1])
+    assert len(df) == 1
+    result = df.with_columns(nw.col("a") + 1, d=nw.col("a") - 1)
+    assert result.columns == ["a", "b", "z", "d"]
+    expected = {"a": [2], "b": [4], "z": [7.0], "d": [0]}
+    compare_dicts(result, expected)
